requirements_docs.txt
sphinx == 3.3.1
numpydoc == 1.1.0
sphinx-rtd-theme == 0.5.0